Combined Tests into single file
