+++ Assets/Scripts/PlayModeTest/CoreGameplayPlayMode.cs
@@ -13,7 +13,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
 
     // ==== Note Bar ====
 
-    /* This test checks that the accuracy text exists within the scene upon the user inputting an action. */
+    /* Test 102: This test checks that the accuracy text exists within the scene upon the user inputting an action. */
     [UnityTest]
     public IEnumerator NoteBarAccuracyTextExists()
     {
@@ -57,7 +57,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         yield return null;
     }
 
-    /* This test checks that the song's physical note objects are instantiated in the proper position. */
+    /* Test 103: This test checks that the song's physical note objects are instantiated in the proper position. */
     [UnityTest]
     public IEnumerator NoteBarNoteSpawnPosition()
     {
@@ -88,7 +88,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the song's physical note objects are instantiated in the proper position. */
+    /* Test 104: This test checks that the song's physical note objects are instantiated in the proper position. */
     [UnityTest]
     public IEnumerator NoteBarNoteSpawnTiming()
     {
@@ -140,7 +140,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the song's physical note objects are destroyed at the proper position. */
+    /* Test 105: This test checks that the song's physical note objects are destroyed at the proper position. */
     [UnityTest]
     public IEnumerator NoteBarNoteDestroyPosition()
     {
@@ -186,7 +186,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
 
     // ==== Characters ====
 
-    /* This test checks that the character's shield spawns as a result of the user selecting 'DEF' from the menu. */
+    /* Test 106: This test checks that the character's shield spawns as a result of the user selecting 'DEF' from the menu. */
     [UnityTest]
     public IEnumerator CharacterSpawnShield()
     {
@@ -221,8 +221,8 @@ public class CoreGameplayPlayMode : MonoBehaviour
 
         Assert.Fail();
     }
-    
-    /* This test checks that the character's shield plays the correct SFX. */
+
+    /* Test 107: This test checks that the character's shield plays the correct SFX. */
     [UnityTest]
     public IEnumerator CharacterShieldSFX()
     {
@@ -255,7 +255,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the character's attack animation plays the correct SFX. */
+    /* Test 108: This test checks that the character's attack animation plays the correct SFX. */
     [UnityTest]
     public IEnumerator CharacterAttackSFX()
     {
@@ -283,11 +283,15 @@ public class CoreGameplayPlayMode : MonoBehaviour
 
     // ==== Menu ====
 
+        // N/A
+
     // ==== Boss ====
+
+        // N/A
 
     // ==== Scene ====
 
-    /* This test checks that the curtain's opening animation plays correctly with proper timing. */
+    /* Test 109: This test checks that the curtain's opening animation plays correctly with proper timing. */
     [UnityTest]
     public IEnumerator SceneCurtainOpeningAnimation()
     {
@@ -309,7 +313,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the spotlights are destroyed upon the curtain's opening animation playing. */
+    /* Test 110: This test checks that the spotlights are destroyed upon the curtain's opening animation playing. */
     [UnityTest]
     public IEnumerator SceneSpotlightDestroy()
     {
@@ -333,7 +337,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the curtains renderer is disabled upon the curtain's opening animation finishing. */
+    /* Test 111: This test checks that the curtains renderer is disabled upon the curtain's opening animation finishing. */
     [UnityTest]
     public IEnumerator SceneCurtainDisableAfterAnimation()
     {
@@ -364,7 +368,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the scene plays the correct applause SFX upon loading in. */
+    /* Test 112: This test checks that the scene plays the correct applause SFX upon loading in. */
     [UnityTest]
     public IEnumerator SceneCurtainApplauseSFX()
     {
@@ -381,7 +385,7 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /* This test checks that the scene plays the correct stage music upon the curtains disabling. */
+    /* Test 113: This test checks that the scene plays the correct stage music upon the curtains disabling. */
     [UnityTest]
     public IEnumerator SceneStageSongPlays()
     {
@@ -407,29 +411,67 @@ public class CoreGameplayPlayMode : MonoBehaviour
         Assert.Fail();
     }
 
-    /*
-    [UnityTest]
-    public IEnumerator SceneSongCompleted()
-    {
-
-    }
-
-    [UnityTest]
-    public IEnumerator SceneSongFailed()
-    {
-
-    }
-
+    /* Test 114: */ 
     [UnityTest]
     public IEnumerator SceneGainCurrency()
     {
+        SetupCoreScene("Assets/Scenes/main.unity");
 
+        yield return null;
+
+        GameObject Menu = GameObject.FindGameObjectWithTag("Menu");
+
+        GlobalLogic logic = new GlobalLogic();
+        GlobalUser user = logic.getGlobalUser();
+
+        int currentCash = user.Cash;
+
+        int count = 0; /* Set the song as finished after 5 seconds, to shorten wait time for the test. */
+        while (Menu.GetComponent<GameLogic>().isSongDone() == false && count < 300)
+        {
+            count++;
+            yield return null;
+        }
+
+        yield return null;
+
+        if (user.Cash > currentCash)
+        {
+            Assert.Pass();
+        }
+
+        Assert.Fail();
     }
 
+    /* Test 115: */
     [UnityTest]
-    public IEnumerator SceneGain()
+    public IEnumerator SceneEXPGain()
     {
+        SetupCoreScene("Assets/Scenes/main.unity");
 
+        yield return null;
+
+        GameObject Menu = GameObject.FindGameObjectWithTag("Menu");
+
+        GlobalLogic logic = new GlobalLogic();
+        GlobalUser user = logic.getGlobalUser();
+
+        int currentXP = user.XP;
+
+        int count = 0; /* Set the song as finished after 5 seconds, to shorten wait time for the test. */
+        while (Menu.GetComponent<GameLogic>().isSongDone() == false && count < 300)
+        {
+            count++;
+            yield return null;
+        }
+
+        yield return null;
+
+        if (user.XP > currentXP)
+        {
+            Assert.Pass();
+        }
+
+        Assert.Fail();
     }
-    */
 }
